requirements.txt
Django=2.2.10
djangorestframework==3.11.0
gunicorn==20.0.4
marshmallow==3.5.1
psycopg2-binary==2.8.4
